refactor(linter): split i18n rules domain (Task 2.3)
- split 391 lines of i18n.rs into logical submodules
- extract hardcode, icon, helpers modules
- establish i18n orchestration module
