+++ crates/katana-linter/src/rules/domains/i18n/mod.rs
@@ -1,0 +1,6 @@
+pub mod hardcode;
+pub mod helpers;
+pub mod icon;
+
+pub use hardcode::lint_i18n;
+pub use icon::lint_icon_facade;
